Fix the Membership_Id newly generated column by EF by changing the datatype to byte like the primary's one

+++ Vidly/Models/Customer.cs
@@ -6,7 +6,7 @@
         public string Name { get; set; }
         public bool IsSubscribedToNewsLetter { get; set; }
         public MembershipType MembershipType { get; set; }
-
-        public int MembershipTypeId { get; set; }
+        //NB: When Using the Foreign key with different datatype ex: int MembershipType entity framework won't consider it as the foriegn key and will generate new One with underscore in the database for applying the 1-m relation
+        public byte MembershipTypeId { get; set; }
     }
 }
